Added DataStore class and implemented object stream
On startup, the program will either create a new DataStore object or retrieve an existing DataStore object from a .dat file. When the program finishes, the DataStore object will be saved to a new .dat file if it does not exist. If the .dat file already exists, the file will be overwritten with the updated DataStore object.