Began implementation of TableView for UserLibraryList
Currently, the TableView is functional for current playlist and my songs. However, I have not figured out how we should handle the list of playlists. I am trying to find out if we can disable the current columns and add new ones with playlist name and date created. If not then I will create another TableView and overlay it on top of the current UserLibraryList.