Added documentation and made UI tweaks
UI Changes:

The search results table is now in a pane with the search bar

The search results are only visible if there is relevant info in the text field. If the search bar is empty, the search results table is invisible.

The all songs search bar no longer affects the my songs table.